+++ Changes.md
@@ -11,6 +11,7 @@ API
 ---
 
 - Menu : Menu items with "checkBox" enabled now draw their checked indicator in place of their "icon" when both are defined.
+- Image : Added `createSwatch()` static method.
 
 1.4.5.0 (relative to 1.4.4.0)
 =======